+++ livekit/livekit/pkg/rtc/room.go
@@ -491,6 +491,7 @@ func (r *Room) Join(
 				p.ToProto(),
 				meta,
 				false,
+				participant.TelemetryGuard(),
 			)
 
 			participant.GetReporter().Tx(func(tx roomobs.ParticipantSessionTx) {
@@ -2028,10 +2029,10 @@ func GetOtherParticipantInfo(
 
 	pInfos := make([]*livekit.ParticipantInfo, 0, len(allParticipants))
 	for _, op := range allParticipants {
-		if !op.Hidden() &&
+		if !(skipSubscriberBroadcast && op.CanSkipBroadcast()) &&
+			!op.Hidden() &&
 			op.Identity() != lpIdentity &&
-			!isMigratingIn &&
-			!(skipSubscriberBroadcast && op.CanSkipBroadcast()) {
+			!isMigratingIn {
 			pInfos = append(pInfos, op.ToProto())
 		}
 	}