+++ livekit/agents-js/plugins/silero/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @livekit/agents-plugin-silero
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`0ea5d771d768bb62ac326b7f0cc580ff1f051c48`](https://github.com/livekit/agents-js/commit/0ea5d771d768bb62ac326b7f0cc580ff1f051c48)]:
+  - @livekit/agents@1.1.0
+
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [[`9dd8641a13cc971866c4683a9122b426927af7c1`](https://github.com/livekit/agents-js/commit/9dd8641a13cc971866c4683a9122b426927af7c1)]:
+  - @livekit/agents@1.0.6
+
 ## 1.0.5
 
 ### Patch Changes